Add Course Delete

+++ teacher/teacher_dashboard.php
@@ -32,7 +32,7 @@
                 $setcos .= "<div class='dropdown_content'>";
                 $setcos .= "<a href='#'>Open</a>";
                 $setcos .= "<a href='teacher_corse_edit.php?course_id={$result_cos['course_id']}'>Edit</a>";
-                $setcos .= "<a href='teacher_course_delete.php?course_id={$result_cos['course_id']}'>Delete</a>";
+                $setcos .= "<a href='teacher_course_delete.php?course_id={$result_cos['course_id']}' onclick=\"return confirm('Are You Sure?🙄')\">Delete</a>";
                 $setcos .= "</div>";
                 $setcos .= "</div>";
